+++ include/madara/expression/SystemCallEval.cpp
@@ -25,7 +25,7 @@ madara::expression::SystemCallEval::~SystemCallEval (void)
 madara::knowledge::KnowledgeRecord
 madara::expression::SystemCallEval::item (void) const
 {
-  return madara::knowledge::KnowledgeRecord::Integer (nodes_.size ());
+  return madara::knowledge::KnowledgeRecord (nodes_.size ());
 }
 
 /// Prune the tree of unnecessary nodes.
